Update the streak counter on rendering a card
The previous implementation updated the streak counter after a card had
been fetched. This made the streak to be off by 1. This was especially
frustrating when the user viewed cards by clicking on them instead of
using the previous/next buttons. In such a case, the streak counter never
updated!