Display the spoiler box for the first card as well
Previously, the spoiler box for the card that starts the queue was
missing. This happened because the dimensions for the card
description area were still zero. The method set up a spoiler box of
height zero. We fixed this by calling the `renderCard` function 2x
whenever we start viewing a deck of cards. This closes #38