Provide feedback for the # search hits
Previously, the user wasn't told whether there were no matching cards or
that they needed to press [Enter] to view all matching cards. This commit
adds such a message.

We also found it worthwhile to put the search bar in its own template so
that we may reuse in the browse page.