Stop overriding the arrow keys
Making it easier to fetch the previous/next card is not enough reason to
remove user ability to keyboard through elements. Sure, we were removing
the override when the user was typing, but that's bug-prone. Navigating
previous/next isn't that bad. We may even create non-conflicting keyboard
shortcuts down the road.

Closes #59