Add the docs changes to master before working on TypeScript (#68)
* Update more docs

* Add a link to Makandra Cards

* Move implementation details of card editing from tutorial

Let the tutorial be accessible to readers of all technical ability. Link
to relevant modules for readers that wish to know more. Furthermore, this
positions the comments close to the source, reducing the chance of them
becoming stale.

* Rebuild the docs

* Move content from main.public_cards.md closer to the source

This is part of an effort to have the docs live as close to the source as
possible.

* Move the wiki material next to the source code.

* Rebuild the docs

* Organize pics on the 'Card Editing' wiki page

* Only display the card + spoiler at the top
* Add a screenshot of tag autocompletion
* Add a screenshot or urgency bars
* Add a screenshot of the shareable URL

* Regenerate the docs without the timestamp

We now exclude the timestamp from being added to the footer. This preserves
doc files that I didn't directly change from getting modified. Before this,
all my docs would get added in one huge commit.