[In Progress] Write the tests' setup and teardown
We renamed `GlobalTestTeardown.ts` to `src/TestConfig.ts` because we do
both the setup and the teardown in that file. Mocha defines setup and
teardown code using hooks, so there's no need for them to be in separate
files.

Before running any test, we ensure that the connection to MongoDB has
been established by waiting for a call from `User.find`. It's amazing
that I use `mongoose.connect` in `MongooseClient.ts` without waiting for
the connection. I guess the timing gods have been merciful to me.

For each test, we first clean the database. We chose to delete all the
documents in each collection instead of dropping the collections
themselves so that we don't have to deal with uninitialized collections
in Mocha.

For the teardown, we still have lots of unhandled promises, and fixing
all of them right now is a pain. We therefore log a shameful reminder
that we need to stop using Mocha's `--exit` workaround.