[In Progress] Add sequelize and postgres deps
I'm rethinking my MongoDB implementation. I need to do more with tags,
but MongoDB is failing tests that I reasonably expect it to pass. Some
of the data is not getting stored when I expect it to. Moving to a
relational database will solve some of my problems, especially through
the use of relations. I want to do more with tagging, and I feel that
relational is the way to go. Right now, I have duplicated content that
I manually keep in sync. That's a sign that I should be using relations.