[Draft] Remove MetadataMongoDB.js
We're doing away with metadata since it complicated the app for no
good reason. I posit that fetching ~200 database rows isn't unbearable.
On the flip side, we don't have to worry about stale data.

Except the functionality that returned `Metadata` documents, the rest of
the functionality has been transferred to other files, e.g. UserModel.ts
and FlashCardModel.ts.

The logic needed to give a snapshot of users' cards, e.g. tag usage and
urgency will be submitted in a subsequent commit.