[Draft] Convert InAppController.js to TypeScript
In addition to converting it, we pruned the object that we pass to the
underlying model. For instance, where we used to pass the entire
`request.body` to create a card, we pass only the fields needed to
create a card. This reduces the chance that we get malicious input since
we only pass properties that SanitizationAndValidation.ts knows how to
deal with.