[Draft] Subsume tests/*. Discontinue Selenium tests.
It's better to have the tests next to the files themselves, e.g. `A.ts`
should have a corresponding `A.Test.ts`. This allows the dev to quickly
see what's being tested and what isn't.

The webdriver tests are too flaky. They do not consistently pass in CI.
We'll go with API tests and mock the app instead of trying to navigate
in Selenium.