[DB] [Review Streaks] Fix validator for last reset time
In d5b4028a15d24038965f2d374d63c5c6b403791a, we naively tried using
Sequelizes `max` validator to assert that the input time is not in the
future. However, that failed, and based on [1], it seems that the only
support is via the `isBefore` API, which is only granular to the day.
This change came in 3172e24a2add92d9e5269a5ae04ab2c42f7059b4, but that
broke the test too.

[1]: https://sequelize.org/master/manual/validations-and-constraints.html#per-attribute-validations