[text-input] Support clearing text
Expose a method, TextInputElement.clearText that clears the value of the input.

Instead of issuing events on keydown, issue them on keyup so that the result of clearing
leaves the text input completely empty. Listening to keydown has the effect of leaving
behind a space, which prevents the placeholder from showing up.